app/requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.20
Werkzeug==1.0.1
alembic==1.4.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
influxdb==5.3.0
itsdangerous==1.1.0
msgpack==0.6.1
paho-mqtt==1.5.1
pip==20.2.4
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
setuptools==50.3.2
six==1.15.0
urllib3==1.25.11